tests have been added

+++ user/src/main/java/com/example/wordstatistic/user/model/Permission.java
@@ -20,7 +20,7 @@ import lombok.*;
 public class Permission {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id") private Long id;
+    @Column(name = "id") private Integer id;
 
     @Column(name = "name", length = 30, nullable = false, unique = true) private String name;
