+++ user/src/main/java/com/example/wordstatistic/user/model/Role.java
@@ -22,7 +22,7 @@ import java.util.Set;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id") private Long id;
+    @Column(name = "id") private Integer id;
 
     @Column(name = "name", length = 20, nullable = false, unique = true) private String name;
