+++ user/src/main/java/com/example/wordstatistic/user/model/User.java
@@ -22,7 +22,7 @@ import java.util.UUID;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id") private Long id;
+    @Column(name = "id") private Integer id;
 
     @Column(name = "uuid", nullable = false, unique = true) private UUID uuid;
